requirements.txt
Django==2.1.5
djangorestframework==3.9.1
feedparser==5.2.1
